Implement tentative boundary
Tentative boundary is a TypeaheadPrediction after which no predictions are displayed.
This way if we are unsure about the prediction we made we can test it before dispaying prediction that were based on this tentative prediction.

The first character prediction for a new TerminalLine is always tentative to detect password prompts.