IJPL-159160 Change meaning of removeBottomEmptyLines method in LinesBuffer
Previously it was taking the `bottomMostLineInd` parameter that was always not less than the index of the last available line. So, this method is needed only to remove the empty lines from the bottom of the buffer. There is no need to allow removing lines in the middle, so it is better to remove `bottomMostLineInd` parameter and assume that we unconditionally remove lines from the bottom of the buffer.