IJPL-159930 Report text buffer width resized event
Width resize requires additional attention because the whole text buffer is rewritten in this operation. Also, some lines might be discarded from history, and it is not tracked by the `linesDiscardedFromHistory` event.
So, for the client it might mean that the whole text buffer is changed, and they have to drop all the caches/recalculate the state and so on.