emulator: fix control sequence parsing of intermediates
Bytes 0x20 to 0x2F are defined as "intermediate" characters. Store these in an new ArrayList instead of as unhandled. When parsing the
sequence used to change the cursor shape ("\x1b[<n> q"), the space between <n> and q is stored as an intermediate, and played back later.
This causes a disappearing character to happen anytime the cursor shape is changed.

Reference: https://vt100.net/emu/dec_ansi_parser
Fixes: #255