emulator: fix control sequence parsing of intermediates (#311)
Bytes 0x20 to 0x2F are defined as "intermediate" characters. Store these in a separate field instead of as unhandled.
Now, when parsing the sequence used to change the cursor shape (`\x1b[<n> q`), the space between `<n>` and `q` is stored as an intermediate. Previously, it was played back; this caused a disappearing character to happen anytime the cursor shape is changed.

Reference:
https://vt100.net/docs/vt510-rm/chapter4.html#S4.3.3
https://vt100.net/docs/vt510-rm/DECSCUSR.html

https://youtrack.jetbrains.com/issue/IJPL-211845

Fixes: #255